Create helpers dir and move menubar js in there
