+++ lib/helpers/menubar.js
