+++ lib/main.js
@@ -1,22 +1,15 @@
 const electron = require('electron')
 const { dialog } = electron
 const realSessions = require('./real-session-data')
-let sessionsArray = realSessions.realSessions
 
-// const menubar = require('menubar')
 const monitor = require('./vendors/active-window')
-const mb = require('./menubar')
+const mb = require('./helpers/menubar')
 
-// TODO resize this for production
-// const mb = menubar({
-//   width: 1200,
-//   height: 1200,
-//   icon: './images/fire.png',
-// })
+const interval = 1000
 
 let activeSession = false
+let sessionsArray = realSessions.realSessions
 let currentSessionName
-const interval = 1000
 let applicationList
 
 mb.on('ready', () => {
